updated ruff to gen fixes, fixed code action generation
