+++ ruffd-core/src/lib.rs
@@ -3,6 +3,7 @@ extern crate lazy_static;
 
 mod notifications;
 mod requests;
+mod ruff_utils;
 pub mod server;
 mod server_ops;
 mod service;
