+++ ruffd-types/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [dependencies]
 lsp-types = "0.93"
-ruff = { git = "https://github.com/charliermarsh/ruff", tag = "v0.0.93", version = "0.0.93" }
+ruff = { git = "https://github.com/charliermarsh/ruff", rev = "f50ff6105637de28f0f51b04c65543c59c90bfaf", version = "0.0.95" }
 tokio = { version = "1.20", features = ["full"] }
 serde = "1.0"
 serde_json = "1.0"
